refactor: line 리스트를 만드는 LineMaker 객체 분리
