+++ src/main/java/domain/Line.java
@@ -1,33 +1,14 @@
 package domain;
 
-import java.util.ArrayList;
 import java.util.List;
 import util.LineGenerator;
+import util.LineMaker;
 
 public class Line {
-
-    private final int numberOfLine;
-    private List<LineStatus> line = new ArrayList<>();
+    private final List<LineStatus> line;
 
     public Line(int numberOfLine, LineGenerator lineGenerator) {
-        this.numberOfLine = numberOfLine;
-        makeLine(lineGenerator);
-    }
-
-    private void makeLine(LineGenerator lineGenerator) {
-        makeFirstLineStatus(lineGenerator);
-        makeElseLineStatus(lineGenerator);
-    }
-
-    private void makeFirstLineStatus(LineGenerator lineGenerator) {
-        line.add(LineStatus.findBy(lineGenerator.generate(false)));
-    }
-
-    private void makeElseLineStatus(LineGenerator lineGenerator) {
-        for (int i = 1; i < this.numberOfLine; i++) {
-            int leftIndex = i - 1;
-            line.add(LineStatus.findBy(lineGenerator.generate(line.get(leftIndex).getStatus())));
-        }
+        this.line = LineMaker.makeLine(lineGenerator, numberOfLine);
     }
 
     public List<LineStatus> getLine() {
