+++ src/test/java/domain/LineTest.java
@@ -19,19 +19,6 @@ public class LineTest {
     }
 
     @Test
-    @DisplayName("라인이 겹치는 경우가 있는지 확인")
-    void makeRandomLine() {
-        int numberOfLine = 1000;
-        Line line = new Line(numberOfLine, new RandomLineGenerator());
-        for (int i = 0; i < numberOfLine - 1; i++) {
-            int rightLine = i+1;
-            if (line.getLine().get(i) == LineStatus.EXIST) {
-                assertThat(line.getLine().get(rightLine).getStatus()).isFalse();
-            }
-        }
-    }
-
-    @Test
     @DisplayName("다리가 존재하는 Line생성")
     void makeExistLine() {
         int numberOfLine = 1;
